--- src/lib/displayAsciiText.ts
@@ -1,9 +1,0 @@
-import figlet from 'figlet';
-
-export const AsciiText = figlet.textSync('GitHub Label Manager', {
-  font: 'Small',
-  horizontalLayout: 'default',
-  verticalLayout: 'default',
-  width: 180,
-  whitespaceBreak: true,
-});
